requirements.txt
# VaaniPariksha - Python Dependencies
# Install with: pip install -r requirements.txt

# Web Framework
flask==3.0.3
flask-cors==4.0.1
flask-sqlalchemy==3.1.1

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.30
alembic==1.13.1

# PDF Processing
pymupdf==1.24.3
reportlab==4.2.0
pillow==10.3.0
pytesseract==0.3.10

# Voice - TTS
pyttsx3==2.90

# Voice - STT (Vosk)
vosk==0.3.45
sounddevice==0.4.6
soundfile==0.12.1

# Security / Encryption
cryptography==42.0.7

# AWS
boto3==1.34.84
botocore==1.34.84

# Utilities
python-dotenv==1.0.1
requests==2.31.0
werkzeug==3.0.3

# NLP (lightweight intent classification & LLM)
scikit-learn==1.4.2
numpy==1.26.4
google-generativeai==0.7.2

# Testing
pytest==8.2.0
pytest-flask==1.3.0
